fix build errors using GCC 13

+++ be/src/format/parquet/parquet_column_convert.h
@@ -407,7 +407,7 @@ public:
                 // half subnormal:
                 //    value = (-1)^sign * (mant / 2^10) * 2^(1 - bias)
                 // half bias = 15 → exponent = 1 - 15 = -14
-                float f = (static_cast<float>(mant) / 1024.0F) * std::powf(2.0F, -14.0F);
+                float f = (static_cast<float>(mant) / 1024.0F) * std::pow(2.0F, -14.0F);
                 return sign ? -f : f;
             }
         } else if (exp == 0x1F) {
